+++ src/sql/util.py
@@ -282,7 +282,7 @@ def find_path_from_root(file_name):
 
         current = current.parent
 
-    return str(Path(current, file_name))
+    return Path(current, file_name)
 
 
 def find_close_match(word, possibilities):
@@ -363,56 +363,85 @@ def parse_toml_error(e, file_path):
         return e
 
 
-def get_user_configs(file_path):
+def get_user_configs(primary_path, alternate_path):
     """
     Returns saved configuration settings in a toml file from given file_path
 
     Parameters
     ----------
-    file_path : str
-        file path to a toml file
+    primary_path : Path
+        file path to toml in project directory
+    alternate_path : Path
+        file path to ~/.jupysql/config
 
     Returns
     -------
     dict
         saved configuration settings
+    Path
+        the path of the file used to get user configurations
     """
-    data = load_toml(file_path)
-    section_names = ["tool", "jupysql", "SqlMagic"]
-    while section_names:
-        section_to_find, sections_from_user = section_names.pop(0), data.keys()
-        if section_to_find not in sections_from_user:
-            close_match = difflib.get_close_matches(section_to_find, sections_from_user)
-            if not close_match:
-                MESSAGE_PREFIX = (
-                    f"Tip: You may define configurations in "
-                    f"{file_path}. Please review our "
-                )
-                display.message_html(
-                    f"{MESSAGE_PREFIX}<a href='{CONFIGURATION_DOCS_STR}'>"
-                    "configuration guideline</a>."
-                )
-                return {}
-            else:
-                raise exceptions.ConfigurationError(
-                    f"{pretty_print(close_match)} is an invalid section "
-                    f"name in {file_path}. "
-                    f"Did you mean '{section_to_find}'?"
-                )
-        data = data[section_to_find]
-    if not data:
-        if section_to_find == "SqlMagic":
-            MESSAGE_PREFIX = (
+    data = None
+    display_tip = True  # Set to true if tip is to be displayed
+    configuration_docs_displayed = False  # To disable showing guidelines once shown
+
+    # Look for user configurations in pyproject.toml and ~/.jupysql/config
+    # in that particular order
+    path_list = [primary_path, alternate_path]
+    for file_path in path_list:
+        section_to_find = None
+        section_found = False
+        if file_path and file_path.exists():
+            data = load_toml(file_path)
+            section_names = ["tool", "jupysql", "SqlMagic"]
+
+            # Look for SqlMagic section in toml file
+            while section_names:
+                section_found = False
+                section_to_find, sections_from_user = section_names.pop(0), data.keys()
+
+                if section_to_find not in sections_from_user:
+                    close_match = difflib.get_close_matches(
+                        section_to_find, sections_from_user
+                    )
+
+                    if not close_match:
+                        if display_tip:
+                            display.message(
+                                f"Tip: You may define configurations in {primary_path}"
+                                f" or {alternate_path}. "
+                            )
+                            display_tip = False
+                        break
+                    else:
+                        raise exceptions.ConfigurationError(
+                            f"{pretty_print(close_match)} is an invalid section "
+                            f"name in {file_path}. "
+                            f"Did you mean '{section_to_find}'?"
+                        )
+
+                section_found = True
+                data = data[section_to_find]
+
+        if section_to_find == "SqlMagic" and section_found and not data:
+            display.message(
                 f"[tool.jupysql.SqlMagic] present in {file_path} but empty. "
-                f"Please review our "
             )
+            display_tip = False
+
+        if not display_tip and not configuration_docs_displayed:
             display.message_html(
-                f"{MESSAGE_PREFIX}<a href='{CONFIGURATION_DOCS_STR}'>"
+                f"Please review our <a href='{CONFIGURATION_DOCS_STR}'>"
                 "configuration guideline</a>."
             )
-    else:
-        display.message(f"Loading configurations from {file_path}")
-    return data
+            configuration_docs_displayed = True
+
+        if not data and not section_found and file_path and file_path.exists():
+            display.message(f"Did not find user configurations in {file_path}.")
+        elif section_found and data:
+            return data, file_path
+
+    return data, None
 
 
 def get_default_configs(sql):