+++ src/tests/test_config.py
@@ -128,6 +128,49 @@ style = "RANDOM"
     assert get_current_configs(magic) == combined
 
 
+def test_load_home_toml_if_sqlmagic_section_not_in_pyproject_toml(
+    tmp_empty, ip_no_magics, capsys, monkeypatch
+):
+    monkeypatch.setattr(
+        Path, "expanduser", lambda path: Path(str(path).replace("~", tmp_empty))
+    )
+    home_toml = Path("~/.jupysql/config").expanduser()
+    home_toml.parent.mkdir(exist_ok=True)
+    home_toml.write_text(
+        """
+[tool.jupysql.SqlMagic]
+autocommit = false
+autolimit = 1
+style = "RANDOM"
+"""
+    )
+
+    Path("pyproject.toml").write_text(
+        """
+[tool.jupysql]
+"""
+    )
+
+    expect = [
+        "Settings changed:",
+        r"autocommit\s*\|\s*False",
+        r"autolimit\s*\|\s*1",
+        r"style\s*\|\s*RANDOM",
+    ]
+
+    config_expected = {"autocommit": False, "autolimit": 1, "style": "RANDOM"}
+
+    os.mkdir("sub")
+    os.chdir("sub")
+
+    load_ipython_extension(ip_no_magics)
+    magic = ip_no_magics.find_magic("sql").__self__
+    combined = {**get_default_testing_configs(magic), **config_expected}
+    out, _ = capsys.readouterr()
+    assert all(re.search(substring, out) for substring in expect)
+    assert get_current_configs(magic) == combined
+
+
 def test_start_ini_default_connection_using_toml_if_any(tmp_empty, ip_no_magics):
     Path("pyproject.toml").write_text(
         """
@@ -284,34 +327,43 @@ style = "RANDOM"
 
 
 @pytest.mark.parametrize(
-    "file_content, param",
+    "file_content, expected_message",
     [
         (
             """
 [tool.jupysql.SqlMagic]
 """,
-            "[tool.jupysql.SqlMagic] present in {path} but empty.",
+            "[tool.jupysql.SqlMagic] present in {primary_path} but empty.",
         ),
-        ("", "Tip: You may define configurations in {path}."),
+        ("", "Tip: You may define configurations in {primary_path} or {alt_path}."),
     ],
     ids=["empty_sqlmagic_key", "missing_sqlmagic_key"],
 )
 def test_loading_toml_display_configuration_docs_link(
-    tmp_empty, ip_no_magics, file_content, param, monkeypatch
+    tmp_empty, capsys, ip_no_magics, file_content, expected_message, monkeypatch
 ):
     Path("pyproject.toml").write_text(file_content)
-    toml_path = str(Path(os.getcwd()).joinpath("pyproject.toml"))
+    toml_path = Path(os.getcwd()).joinpath("pyproject.toml")
+    config_path = Path("~/.jupysql/config").expanduser()
 
     os.mkdir("sub")
     os.chdir("sub")
+
     mock = Mock()
     monkeypatch.setattr(display, "message_html", mock)
-
     load_ipython_extension(ip_no_magics)
+    out, _ = capsys.readouterr()
+
     param = (
-        f"{param.format(path=toml_path)} Please review our "
+        f"Please review our "
         f"<a href='{CONFIGURATION_DOCS_STR}'>configuration guideline</a>."
     )
+
+    expected_message = expected_message.format(
+        primary_path=str(toml_path), alt_path=str(config_path)
+    )
+
+    assert expected_message in out
     mock.assert_called_once_with(param)
 
 
@@ -461,3 +513,114 @@ autocommit = true
         "To load settings from pyproject.toml or ~/.jupysql/config, "
         "install with: pip install toml"
     ) in out
+
+
+@pytest.mark.parametrize(
+    "pyproject_content, config_content, expected_messages",
+    [
+        (
+            "",
+            "",
+            [
+                (
+                    "Tip: You may define configurations in "
+                    "{pyproject_path} or {config_path}."
+                ),
+                "Did not find user configurations in {pyproject_path}.",
+                "Did not find user configurations in {config_path}.",
+            ],
+        ),
+        (
+            "",
+            "[tool.jupysql.SqlMagic]",
+            [
+                (
+                    "Tip: You may define configurations in "
+                    "{pyproject_path} or {config_path}."
+                ),
+                "Did not find user configurations in {pyproject_path}.",
+                "[tool.jupysql.SqlMagic] present in {config_path} but empty.",
+            ],
+        ),
+        (
+            "",
+            """
+[tool.jupysql.SqlMagic]
+feedback=True
+autopandas=True
+""",
+            [
+                (
+                    "Tip: You may define configurations in "
+                    "{pyproject_path} or {config_path}."
+                ),
+                "Did not find user configurations in {pyproject_path}.",
+            ],
+        ),
+        (
+            "[tool.jupysql.SqlMagic]",
+            "",
+            [
+                "[tool.jupysql.SqlMagic] present in {pyproject_path} but empty.",
+                "Did not find user configurations in {config_path}.",
+            ],
+        ),
+        (
+            "[tool.jupysql.SqlMagic]",
+            "[tool.jupysql.SqlMagic]",
+            [
+                "[tool.jupysql.SqlMagic] present in {pyproject_path} but empty.",
+                "[tool.jupysql.SqlMagic] present in {config_path} but empty.",
+            ],
+        ),
+        (
+            "[tool.jupysql.SqlMagic]",
+            """
+[tool.jupysql.SqlMagic]
+feedback=True
+autopandas=True
+""",
+            [
+                "[tool.jupysql.SqlMagic] present in {pyproject_path} but empty.",
+            ],
+        ),
+    ],
+)
+def test_user_config_load_sequence_and_messages(
+    tmp_empty,
+    ip_no_magics,
+    monkeypatch,
+    capsys,
+    pyproject_content,
+    config_content,
+    expected_messages,
+):
+    toml_path = Path("pyproject.toml")
+    toml_path.touch(exist_ok=True)
+    toml_path.write_text(pyproject_content)
+
+    Path("~/.jupysql").expanduser().mkdir(parents=True, exist_ok=True)
+    config_path = Path("~/.jupysql/config").expanduser()
+    config_path.touch(exist_ok=True)
+    config_path.write_text(config_content)
+
+    toml_path = str(Path(os.getcwd()).joinpath("pyproject.toml"))
+    config_path = str(Path("~/.jupysql/config").expanduser())
+
+    mock = Mock()
+    monkeypatch.setattr(display, "message_html", mock)
+    load_ipython_extension(ip_no_magics)
+    out, _ = capsys.readouterr()
+
+    param = (
+        f"Please review our "
+        f"<a href='{CONFIGURATION_DOCS_STR}'>configuration guideline</a>."
+    )
+
+    for message in expected_messages:
+        expected_message = message.format(
+            pyproject_path=str(toml_path), config_path=str(config_path)
+        )
+        assert expected_message in out
+
+    mock.assert_called_once_with(param)